added endpoint for Review entity

+++ src/main/java/com/example/doctor_book_app_backend/entity/Review.java
@@ -27,7 +27,7 @@ public class Review {
     @Column(name = "date")
     private Date date;
     @Column(name = "rating")
-    private int rating;
+    private Float rating;
     @Column(name = "comment")
     private String comment;
 }
